Provide an initial module system
by expanding the two UIs (gui, cli).

The module system suggested is to split the modules into gui, which carries over Qt and attaches menus and toolbars to the main application, cli, which attaches parsers to the cli and core, which contains functionality used by both. CMakeLists have been modified to always load core, and depending on the build, load cli or gui.

The implementation aims to keep depthmapX agnostic to the modules which will be provided in the modules folder according to the example, and by providing a self-registering system.

For this first implementation an expansion of the gui is provided by expanding the MainWindow and adding menu items. The example module will create shortest paths between two segments